store_image_function/requirements.txt
functions-framework==3.5.0
google-cloud-storage

